requirements.txt
python-docx>=1.1.0



numpy~=2.0.1
pillow~=11.1.0
onnxruntime~=1.22.1
mysql-connector-python~=9.3.0
pandas~=2.3.3
pymysql~=1.1.1
openpyxl~=3.1.5
bcrypt~=4.3.0